// YerbowoFrontend/src/app/account/models/addressEdit.ts
export interface AddressEdit {
    id: number;
    userId: number;
    alias: string;
    firstName: string;
    lastName: string;
    street: string;
    buildingNumber: string;
    apartmentNumber: string;
    place: string;
    postCode: string;
    phone: string;
    email: string;
    nip: string;
    company: string;
}